compute-service/requirements.txt
fastapi==0.116.1
httpx==0.28.1
Pillow==11.3.0
numpy==2.3.2
matplotlib==3.10.5
boto3==1.40.37
botocore==1.40.37
python-dotenv==1.1.1
pydantic==2.11.7
uvicorn==0.35.0
